+++ src/check/mod.rs
@@ -1,6 +1,16 @@
+mod dir_1;
+mod dir_2;
+mod doc_1;
+mod doc_2;
 pub(crate) mod issue;
 mod location;
 pub(crate) mod scanners;
+mod state_1;
 
+pub(crate) use dir_1::check_dir_1;
+pub(crate) use dir_2::check_dir_2;
+pub(crate) use doc_1::check_doc_1;
+pub(crate) use doc_2::check_doc_2;
 pub(crate) use issue::Issue;
 pub(crate) use location::Location;
+pub(crate) use state_1::State1;
