+++ src/check/state_1.rs
@@ -1,0 +1,11 @@
+use super::Issue;
+use ahash::AHashMap;
+use std::path::PathBuf;
+
+/// mutable state accumulated in check phase 1
+pub struct State1 {
+    pub issues: Vec<Issue>,
+    pub linked_resources: Vec<PathBuf>,
+    pub title_variants: AHashMap<String, u32>,
+    pub level_variants: AHashMap<String, AHashMap<u8, u32>>,
+}
